Move the trace module
